rename to glucose -model -controller -router to match frontend
